promtail: restore build support for non-linux platforms
Support for building promtail on non-Linux platforms (and Linux
platforms where CGO has been disabled) has been added.

This commit also reverts accidentally commited code in journaltarget.go
that panics if a timestamp occurs before the timestamp of the previous
received journal entry.
