+++ pkg/promtail/targets/journaltarget_test.go
@@ -1,4 +1,4 @@
-// +build linux
+// +build linux,cgo
 
 package targets
